+++ CHANGELOG.md
@@ -1,7 +1,13 @@
+# Version 1.0.2
+
+- Add event emitter to allow other extensions to listen to events
+	- 'sql-query:request' - Fired before variables are replaced in the query
+	- 'sql-query:response' - Fired after the query has been executed
+
 # Version 1.0.1
 
 - Update sumary row basedo on searched results
 
 # Version 1.0.0
 
-First version
+First release